Reduced CPU starvation in tests for more unaffected results Changed type id type from int to byte for less footprint That's all shows around 15-20% performance increasing (but actually it's around 5-8% with avoiding vtable lookups for types). And the rest is an invisible performance, that was hidden in tests but actually didn't slow down applications.

+++ src/main/java/com/github/mangelion/ClickHousePacketDecoder.java
@@ -62,7 +62,7 @@ final class ClickHousePacketDecoder extends ReplayingDecoder<Void> {
         ColumnWithTypeAndName[] cs = new ColumnWithTypeAndName[columns];
         for (int i = 0; i < columns; i++) {
             String columnName = readStringBinary(in);
-            int type = ColumnType.valueOf(readStringBinary(in));
+            byte type = ColumnType.valueOf(readStringBinary(in));
 
             ColumnWithTypeAndName column = cs[i] = new ColumnWithTypeAndName(type, columnName, ctx.alloc().directBuffer());
